% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centers}
\alias{centers}
\title{Centroid Information for Hex Map}
\format{
A \code{data.table} with 51 rows and 3 variables:
\describe{
  \item{x, y}{Centroid information}
  \item{id}{Two-letter state code}
}
}
\usage{
centers
}
\description{
A dataset with locations of the centers of the hex grid hexes
}
\keyword{datasets}
