% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_pop}
\alias{county_pop}
\title{County Population}
\format{
A \code{data.table} with 3143 rows and 35 variables.
}
\source{
\url{US Census}
}
\usage{
county_pop
}
\description{
A dataset with county-level information from the 2020-2021 census estimates. See \url{https://www2.census.gov/programs-surveys/popest/technical-documentation/file-layouts/2020-2021/CO-EST2021-ALLDATA.pdf} for variable definitions.
}
\keyword{datasets}
