% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{dodged_labeled_bar}
\alias{dodged_labeled_bar}
\title{Grouped bar graph with labeled values on top}
\usage{
dodged_labeled_bar(
  dat,
  xvar,
  yvar,
  fillvar,
  yscale = scales::label_number(big.mark = ","),
  limits = NULL
)
}
\arguments{
\item{dat}{Data.}

\item{xvar}{x-axis variable, as a string.}

\item{yvar}{y-axis variable, as a string.}

\item{fillvar}{Group variable for the bars}

\item{yscale}{Numerical formatting function, in \code{scales::label_} style.}

\item{limits}{y-axis limits on the graph, usually to make it taller so as to fit the text.}
}
\description{
Given an explicit bar height, constructs a bar graph with multiple categories and puts the value given for the height on top of the bars, formatted, in Serif size 13 font.
}
\examples{

# Example data
data.frame(x = c('A','B', 'A','B'), group = c('C','C','D','D'), y = c(4, 5,6,7)) |>
    dodged_labeled_bar('x','y', 'group')

}
