% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{dynamite_plot}
\alias{dynamite_plot}
\title{Aggregates to a mean value and plots that value, with 95 percent confidence interval}
\usage{
dynamite_plot(
  dat_orig,
  xvar,
  yvar,
  yscale = scales::label_number(big.mark = ","),
  fillcolor = "lightblue",
  highlight = NULL,
  highlightcolor = "firebrick",
  limits = NULL,
  cis = TRUE,
  hjust = ifelse(cis == TRUE, 1.5, 0.5),
  savedata = NULL
)
}
\arguments{
\item{xvar}{x-axis variable, as a string.}

\item{yvar}{y-axis variable to aggregate, as a string.}

\item{yscale}{Numerical formatting function, in \code{scales::label} style.}

\item{fillcolor}{The color of the bars.}

\item{highlight}{Set of categories to highlight}

\item{highlightcolor}{The color of the highlighted bars}

\item{limits}{y-axis limits on the graph, usually to make it taller so as to fit the text.}

\item{cis}{Show confidence intervals}

\item{hjust}{Horizontal justification for the mean labels}

\item{savedata}{Save the graphed data to CSV file by this name.}

\item{dat}{Data.}
}
\description{
Given a grouping variable and a value to aggregate, gets the mean and 95% CI, and plots it on a labeled bar graph with a confidence interval
}
\examples{

# Not now

}
