% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{get_thetext}
\alias{get_thetext}
\title{Creates a tokenized data set, including single words and pre-specified two-word phrases}
\usage{
get_thetext(
  dat_orig,
  textcol,
  keepvars = NULL,
  stopwords = NULL,
  hardcode = NULL,
  tolower = TRUE
)
}
\arguments{
\item{dat_orig}{The data}

\item{textcol}{String for the column name with text to tokenize}

\item{keepvars}{String vector of columns to keep with the data. Note an "ID" column will also be added.}

\item{stopwords}{String vector of words to remove}

\item{hardcode}{String vector of phrases to detect (more than one word; all one-word phrases will already be detected). Not case sensitive.}

\item{tolower}{Convert all text to lowercase?}
}
\description{
Creates a tokenized data set, including single words and pre-specified two-word phrases
}
