% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{grouped_colorlab_bar}
\alias{grouped_colorlab_bar}
\title{Grouped dynamite plot with median markers and group labels at the top}
\usage{
grouped_colorlab_bar(
  dat_orig,
  xvar,
  catvar,
  yvar,
  yscale = scales::label_number(big.mark = ","),
  catcolor = NULL,
  limits = NULL,
  cis = TRUE,
  medians = TRUE,
  hjust = ifelse(cis == TRUE, 1.5, 0.5),
  lpos = 0.5,
  lcolor = "white",
  lfamily = "serif",
  lsize = 14/ggplot2::.pt,
  tfamily = NULL,
  tsize = 14,
  help_ordering = TRUE,
  savedata = NULL
)
}
\arguments{
\item{xvar}{x-axis variable, as a string.}

\item{catvar}{Variable indicating the categories that each of the x-axis values slot into.}

\item{yvar}{y-axis variable to aggregate, as a string.}

\item{yscale}{Numerical formatting function, in \code{scales::label} style.}

\item{catcolor}{A labeled (or properly ordered) vector connecting each value of \code{catvar} to a color.}

\item{limits}{y-axis limits on the graph, usually to make it taller so as to fit the text.}

\item{cis}{Show confidence intervals}

\item{medians}{Show medians}

\item{hjust}{Horizontal justification for mean text labels}

\item{lpos}{Y-value height of the mean text labels}

\item{lcolor}{Color for mean text labels}

\item{lfamily}{Font family for mean text labels}

\item{lsize}{Font size for mean text labels}

\item{tfamily}{Font family for theme (will use \code{showtext_font} in \code{theme_nick()})}

\item{tsize}{Font size for theme}

\item{help_ordering}{Reorder the levels of the \code{xvar} variable so that they're ordered within values of \code{catvar}.}

\item{savedata}{Save the graphed data to CSV file by this name.}

\item{dat}{Data.}
}
\description{
Given a grouping variable and a value to aggregate, gets the mean and 95% CI, and plots it on a labeled bar graph with a confidence interval
}
\details{
To control the ordering of the bars, set \code{help_ordering} to \code{FALSE}, and make sure that both \code{xvar} and \code{catvar} are ordered such that the \code{xvar} values are ordered such that all the same \code{catvar}s go together.
}
\examples{

# Not now

}
