% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{labeled_bar}
\alias{labeled_bar}
\title{Bar graph with labeled values on top}
\usage{
labeled_bar(
  dat,
  xvar,
  yvar,
  yscale = scales::label_number(big.mark = ","),
  fillcolor = "lightblue",
  limits = NULL
)
}
\arguments{
\item{dat}{Data.}

\item{xvar}{x-axis variable, as a string.}

\item{yvar}{y-axis variable, as a string.}

\item{yscale}{Numerical formatting function, in \code{scales::label_} style.}

\item{fillcolor}{The color of the bars.}

\item{limits}{y-axis limits on the graph, usually to make it taller so as to fit the text.}
}
\description{
Given an explicit bar height, constructs a bar graph and puts the value given for the height on top of the bar, formatted, in Serif size 13 font.
}
\examples{

# Example data
data.frame(x = c('A','B'), y = c(4, 5)) |>
    labeled_bar('x','y')

}
