% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{line_w_ribbon}
\alias{line_w_ribbon}
\title{Line Graph with Confidence Interval Ribbon}
\usage{
line_w_ribbon(
  dat_orig,
  xvar,
  yvar,
  colorvar = NULL,
  yscale = scales::label_number(),
  limits = NULL,
  savedata = NULL
)
}
\arguments{
\item{dat_orig}{Data.}

\item{xvar}{x-axis variable, as a string.}

\item{yvar}{y-axis variable, as a string.}

\item{colorvar}{Color variable to do multiple lines at once}

\item{yscale}{Numerical formatting function, in \code{scales::label_} style.}

\item{limits}{y-axis limits on the graph, usually to make it taller so as to fit the text.}

\item{savedata}{Save the graphed data to a CSV file by this name.}
}
\description{
Calculates the average of \code{yvar} as well as a 95% confidence interval for that mean, within each \code{xvar} value. Then produces a line graph with a confidence ribbon. Will drop any observation for which the confidence interval has a range of 0.
}
\examples{

# Example data
line_w_ribbon(mtcars, 'cyl', 'mpg')

}
