% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_functions.R
\name{linear_slope}
\alias{linear_slope}
\title{Get the linear slope between two variables}
\usage{
linear_slope(
  dat = NULL,
  formula = NULL,
  y = NULL,
  x = NULL,
  se = FALSE,
  p = FALSE,
  names = (se | p),
  ...
)
}
\arguments{
\item{formula}{Formula}

\item{y}{Instead pass the dependent variable as a vector (only works with a single predictor).}

\item{x}{Instead pass a single predictor as a vector.}

\item{se}{Instead of just returning the coefficient, pass a vector that also includes the standard error.}

\item{p}{Instead of just returning the coefficient, pass a vector that also includes the p-value (will be third element if \code{se = TRUE}).}

\item{names}{Name the return vector}

\item{...}{Other arguments to pass to \code{fixest::feols}}

\item{data}{Dataset}
}
\description{
This performs a linear regression using \code{fixest::feols} and returns only the second coefficient (i.e. the first non-intercept coefficient).
}
\examples{

# Example data
eval(pp('2','+','2'))

}
