% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_functions.R
\name{mean_rm}
\alias{mean_rm}
\title{Mean with na.rm}
\usage{
mean_rm(...)
}
\arguments{
\item{...}{To pass to \code{mean()}.}
}
\description{
This is just a mean function but with na.rm set to TRUE.
}
\examples{

# Example data
mean_rm(c(1, NA))

}
