% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicksshorts-package.R
\docType{package}
\name{nicksshorts-package}
\alias{nicksshorts}
\alias{nicksshorts-package}
\title{nicksshorts: Some Shortcut Functions for Nick's Analysis}
\description{
Just some neat shortcut functions I shall use.
}
\keyword{internal}
