% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{pdv}
\alias{pdv}
\title{Calculates the z score of the test that two proportions are the same}
\usage{
pdv(p1, p2, n1, n2)
}
\arguments{
\item{p1, p2}{Proportion in first/second group.}

\item{n1, n2}{Sample size in first/second group}
}
\description{
Arguments \code{p1} and \code{p2} only are vectorized. This is intended for use when you have a lot of different proportions to test (such as proportion of a word present) and the sample sizes are constant across all tests.
}
