% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_functions.R
\name{pp}
\alias{pp}
\title{Convenience function for paste0 and parse}
\usage{
pp(...)
}
\arguments{
\item{...}{To pass to \code{paste0()}.}
}
\description{
This is just \code{parse(text = paste0(...))} to be fed to \code{eval}.
}
\examples{

# Example data
eval(pp('2','+','2'))

}
