% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_functions.R
\name{qrecode}
\alias{qrecode}
\title{Fast-to-type recoding function}
\usage{
qrecode(dat, var, from, to, newvar = var, checkfrom = FALSE, difftype = FALSE)
}
\arguments{
\item{dat}{Original \code{data.table}. Note this data will be directly modified in-place.}

\item{var}{String name of the variable you want to recode.}

\item{from}{Vector of values to recode from. Any value of \code{var} left out here will be maintained as-is, unless \code{difftype = TRUE}.}

\item{to}{Vector of values of the same length as \code{from} to recode to, in the same order.}

\item{newvar}{If you want the recoded values in another variable, give the string name of that new variable here.}

\item{checkfrom}{Check that all values of \code{from} are actually in the data, and give a warning if they're not. This will make the code run slower.}

\item{difftype}{Instead of starting by copying all the old \code{var} values, start with a \code{NULL} column. This will allow the recoded values to be a different type from the original. This requires that \code{newvar} not be the same as \code{var}.}
}
\description{
Recodes just some of the values in one variable into another.
}
\examples{

dat = data.table(myvar = c('a','b','c'))
qrecode(dat, 'myvar',c('a','c'),c('d','e'))
}
