% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_functions.R
\name{quantile_rm}
\alias{quantile_rm}
\title{Quantile function with na.rm}
\usage{
quantile_rm(..., prob = 0.5)
}
\arguments{
\item{...}{To pass to \code{quantile()}.}

\item{prob}{The quantile to get, by default the median.}
}
\description{
This is just a quantile function, by default the median, but with na.rm set to TRUE.
}
\examples{

# Example data
quantile_rm(c(1, 2, NA))

}
