% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{rangescale}
\alias{rangescale}
\alias{label_rangescale}
\title{Numeric Labeling Across a Wide Range}
\usage{
rangescale(
  x,
  accuracy = 1,
  prefix = "",
  big.mark = ",",
  decimal.mark = ".",
  trim = TRUE,
  ...
)

label_rangescale(
  accuracy = 1,
  prefix = "",
  big.mark = ",",
  decimal.mark = ".",
  trim = TRUE,
  ...
)
}
\arguments{
\item{x}{Data.}

\item{accuracy}{For the most part, see \code{scales::comma}. However, this also accepts a vector of options. If a vector is set, it will apply that value to the sub-1000, 1k-1M, 1M-1B, 1B-1T options, in that order. If fewer than five options are set, the last option will be copied forward (so \code{c(.1, 1)} will give an accuracy of .1 to sub-1k and 1 to the rest).}

\item{prefix, big.mark, decimal.mark, trim, ...}{See \code{scales::comma}. Note suffix and scale cannot be set, and accuracy default is now 1.}
}
\description{
Applies scales::comma labeling. Chooses appropriate scaling and adds a suffix separately for sub-1000, 1k-1M, 1M-1B, 1B-1T, 1T+ Above 1T keeps reporting trillions.
}
\details{
\code{label_rangescale()} returns a version of the \code{rangescale} function with the appropriate options selected, like the other \code{label_} functions in the scales package.
}
\examples{

# Example data
rangescale(c(400, 4000, 4200000, 4300000000))
rangescale(c(400, 4000, 4200000, 4300000000), accuracy = c(1, .1))
my_range_fcn = label_rangescale(accuracy = c(1, .1))
my_range_fcn(c(400, 4000, 4200000, 4300000000))

}
