% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_functions.R
\name{sd_rm}
\alias{sd_rm}
\title{Standard deviation with na.rm}
\usage{
sd_rm(...)
}
\arguments{
\item{...}{To pass to \code{sd()}.}
}
\description{
This is just a sd function but with na.rm set to TRUE.
}
\examples{

# Example data
sd_rm(c(1, 2, NA))

}
