% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spdf_fortified}
\alias{spdf_fortified}
\title{State Hex Map}
\format{
A \code{data.table} with 357 rows and 7 variables:
\describe{
  \item{long, lat, order, hold, piece, group}{Vertex information}
  \item{id}{Two-letter state code}
}
}
\usage{
spdf_fortified
}
\description{
A dataset with a hexgrid map of the United States
}
\keyword{datasets}
