% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_pop}
\alias{state_pop}
\title{State Population}
\format{
A \code{data.table} with 51 rows and 33 variables
}
\source{
\url{US Census}
}
\usage{
state_pop
}
\description{
A dataset with state-level information from the 2020-2021 census estimates, including population. See \url{https://www2.census.gov/programs-surveys/popest/technical-documentation/file-layouts/2020-2021/CO-EST2021-ALLDATA.pdf} for variable definitions.
}
\keyword{datasets}
