% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_functions.R
\name{sum_rm}
\alias{sum_rm}
\title{Sum with na.rm}
\usage{
sum_rm(...)
}
\arguments{
\item{...}{To pass to \code{sum()}.}
}
\description{
This is just a sum function but with na.rm set to TRUE.
}
\examples{

# Example data
sum_rm(c(1, NA))

}
