% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{theme_nick}
\alias{theme_nick}
\title{A Theme I Like}
\usage{
theme_nick(size = 14, showtext_font = NULL, ...)
}
\arguments{
\item{size}{Text font size}

\item{showtext_font}{String name of font to be included. Assumes that \code{showtext::showtext_auto()} has been set beforehand. Will set \code{lineheight} to .5 by default to offset \code{showtext} oddness.}

\item{...}{To pass to \code{ggpubr::theme_pubr()}.}
}
\description{
This combines \code{ggpubr::theme_pubr()} with serif text and a horizontal y-axis title. Don't forget to add a \code{scales::label_wrap()} for long y-axis titles!
}
\examples{

# Example data
ggplot(mtcars, aes(x = mpg, y = hp)) + theme_nick()

}
