% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{topn_table}
\alias{topn_table}
\title{Sorted, formatted table with formatted values and top-n selection}
\usage{
topn_table(
  dat_orig,
  ntop = min(10, nrow(dat)),
  vars = names(dat),
  sort = tail(vars, 1),
  order = -1L,
  vscales = NULL,
  rank = TRUE,
  savedata = NULL
)
}
\arguments{
\item{dat_orig}{Data.}

\item{ntop}{How many rows to keep}

\item{vars}{Variables to keep in the table}

\item{sort}{Which variable to sort by}

\item{order}{Sort order (-1 = descending)}

\item{vscales}{List of \code{scales::label_} formatting functions to format the columns of data for presentation, or \code{NULL} to just present raw values.}

\item{rank}{Add a "Rank" variable showing the row numbers to the left of the table.}

\item{savedata}{Save the full data (before trimming to top N) to cSV file by this name}
}
\description{
Sorts a table by a value, picks the top N values, and then produces a nice looking table.
}
\examples{

# Example data
data.frame(x = c('A','B', 'A','B'), group = c('C','C','D','D'), y = c(4, 5,6,7)) |>
    topn_table(2)

}
