% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{us_hex_map}
\alias{us_hex_map}
\title{Hexgrid US State Map}
\usage{
us_hex_map(
  dat_orig,
  stvar,
  yvar,
  yscale = scales::label_number(big.mark = ","),
  na.include = TRUE,
  grad_low = "#56B1F7",
  grad_mid = NULL,
  grad_high = "#C0C0C0"
)
}
\arguments{
\item{dat_orig}{Data}

\item{stvar}{Column name with the two-letter state code in it.}

\item{yvar}{Variable to determine color intensity.}

\item{yscale}{Function}

\item{na.include}{Include states not listed in \code{dat_orig} as blanks.}

\item{grad_low, grad_mid, grad_high}{Colors to use for gradient intensity. \code{grad_mid} is optional.}
}
\description{
Creates a hexgrid map of the United States. CURRENTLY NOT WORKING DUE TO RETIREMENT OF RGEOS AND RGDAL
}
