% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_functions.R
\name{weighted_mean_rm}
\alias{weighted_mean_rm}
\title{Weighted mean function with na.rm}
\usage{
weighted_mean_rm(x, w, ...)
}
\arguments{
\item{x}{Value to take the mean of}

\item{w}{Weights}

\item{...}{Other arguments to pass along}
}
\description{
This is just a weighted mean function, but with \code{na.rm = TRUE} (and adjusted so that it actually works).
}
\examples{

# Example data
weighted_mean_rm(c(NA,1,2),c(1,NA,2))

}
