% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_functions.R
\name{weighted_median_rm}
\alias{weighted_median_rm}
\title{Weighted median function with na.rm}
\usage{
weighted_median_rm(x, w)
}
\arguments{
\item{x}{Value to take the median of}

\item{w}{Weights}
}
\description{
This is a weighted median function, which returns the value X for which 50% of the weight is for values X or lower, and less than 50% of the weight is for values lower than X. Missing values are always dropped.
}
\details{
Should work with any sortable \code{x} variable, including strings.
}
\examples{

# Example data
weighted_median_rm(c(NA,1,2,3,4),c(1,NA,2,3,10))

}
